requirements.txt
pandas
numpy
scikit-learn
Pillow
ImageHash
plotly
pandarallel
pyarrow
ipykernel
tqdm
opencv-python
matplotlib
ipywidgets
nbformat
lightgbm
nb-clean